refactor: RSS package
